Refactor module structure
