+++ crates/hir-analysis/src/lib.rs
@@ -57,19 +57,18 @@ pub struct Jar(
     ty::trait_def::impls_of_trait,
     ty::trait_def::impls_of_ty,
     // Trait constraints
-    ty::constraint::collect_super_traits,
-    ty::constraint::collect_trait_constraints,
-    ty::constraint::collect_adt_constraints,
-    ty::constraint::collect_implementor_constraints,
-    ty::constraint::collect_impl_block_constraints,
-    ty::constraint::collect_func_def_constraints,
-    ty::constraint::collect_func_def_constraints_impl,
-    ty::constraint::ty_constraints,
-    ty::constraint_solver::check_ty_wf,
-    ty::constraint_solver::check_trait_inst_wf,
-    // Trait Resolution,
+    ty::trait_resolution::constraint::collect_super_traits,
+    ty::trait_resolution::constraint::collect_trait_constraints,
+    ty::trait_resolution::constraint::collect_adt_constraints,
+    ty::trait_resolution::constraint::collect_implementor_constraints,
+    ty::trait_resolution::constraint::collect_impl_block_constraints,
+    ty::trait_resolution::constraint::collect_func_def_constraints,
+    ty::trait_resolution::constraint::collect_func_def_constraints_impl,
+    ty::trait_resolution::constraint::ty_constraints,
     ty::trait_resolution::PredicateListId,
     ty::trait_resolution::is_goal_satisfiable,
+    ty::trait_resolution::check_ty_wf,
+    ty::trait_resolution::check_trait_inst_wf,
     // Type checking.
     ty::ty_check::check_func_body,
     // Diagnostic accumulators.
